third_party/WebKit/LayoutTests/inspector/sources/debugger/script-formatter-breakpoints-1-expected.txt

Tests the script formatting is working fine with breakpoints.

Debugger was enabled.

Running: testBreakpointsInOriginalAndFormattedSource
Script execution paused.
Breakpoint sidebar pane while paused in raw

    script-formatter-breakpoints-1.html:12var f=0;
Script execution resumed.
Script execution paused.
Breakpoint sidebar pane while paused in pretty printed

    script-formatter-breakpoints-1.html:formatted:17var f = 0;
Breakpoint sidebar pane while paused in raw

    script-formatter-breakpoints-1.html:12var f=0;
Script execution resumed.
Debugger was disabled.
